requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
firebase-admin
requests